Move PropertiesApplier to common project
